simple hs integration (demonstrator)
This is a straightforward patch of Stack.Setup to call out to `hs`
to discover GHC toolchain installations that may reside in `stack`
or `ghcup` for Unix (some adjustments will need to be made for
Windows).

Done properly, this behaviour will be controlled by a global
configuration of course -- this is just a preview/demonstrator.

The version has been bump to 2.8.0.1.